% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Perm_ID.R
\name{add_Perm_ID}
\alias{add_Perm_ID}
\title{Pulls out and combine permutations from the permutation dataframe made by the permutations function by Rsample.}
\usage{
add_Perm_ID(perm_dat)
}
\arguments{
\item{data}{The permutations data frame}
}
\value{
A dataframe the permuted data with permutation ID labels
}
\description{
Pulls out and combine permutations from the permutation dataframe made by the permutations function by Rsample.
}
