% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvals.R
\name{calculate_pvals}
\alias{calculate_pvals}
\title{Calculates the p-values for the permuted data.frame.}
\usage{
calculate_pvals(data, taxon, n)
}
\arguments{
\item{data}{The data frame}

\item{n}{the number of permutations}

\item{class}{The taxonomic level being tested on}
}
\value{
A data.frame containing p-values for the permuted data arranged in ascending order by p-val
}
\description{
Calculates the p-values for the permuted data.frame.
}
