% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differences.R
\name{get_differences}
\alias{get_differences}
\title{Calculates the differences between two columns.}
\usage{
get_differences(data, col1, col2)
}
\arguments{
\item{data}{The data frame}

\item{col1}{the column to be subtracted from}

\item{col2}{the column to be subtracted by
the function will do col1 - col2}
}
\value{
A mutated dataframe containing the differences between column i and l
}
\description{
Calculates the differences between two columns.
}
