% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_difference_dt.R
\name{get_differences_dt}
\alias{get_differences_dt}
\title{Calculates the differences between two columns with data.table.}
\usage{
get_differences_dt(data, col1, col2)
}
\arguments{
\item{data}{The data frame}

\item{col1}{the column to be subtracted from}

\item{col2}{the column to be subtracted by
the function will do col1 - col2}
}
\value{
A data.table containing the differences between column i and l
}
\description{
Calculates the differences between two columns with data.table.
}
