% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perm_pvals.R
\name{get_perm_pvals}
\alias{get_perm_pvals}
\title{Master function that gets p-values for permuted data.}
\usage{
get_perm_pvals(data, taxon, n, ...)
}
\arguments{
\item{data}{The data frame}

\item{taxon}{The taxonomic level being tested on}

\item{n}{the number of permutations}

\item{...}{vector of treatments to be compared to one another}
}
\value{
A dataframe containing p-values for the permuted data arranged in ascending order by p-val
}
\description{
Master function that gets p-values for permuted data.
}
